requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
requests==2.32.3
scapy==2.6.1
six==1.17.0
urllib3==2.3.0
